requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-mistralai
langchain-chroma
chromadb
faster-whisper
pytest
transformers
Pillow
torch

